+++ CHANGELOG.md
@@ -18,6 +18,10 @@
     - Features:
       - CLASS1.INFORMATION Type=69 - 76 added.
       - CLASS1.CONTROL Type=42 - 43 added.
+      - VSCP_TIMER_ID_INVALID introduced for invalid timer id. Please update your code accordingly, see template 'vscp_timer.c'.
+      - The internal VSCP core time (time since epoch) is now updated once per second.
+        The core process routine will handle it, independent of the current internal state.
+        Please increase your max. number of timers to be able to use this feature.
 
 ## 0.6.1